refactor(LottoIssuer): 로또를 발행하는 로직을 분리해 새로운 클래스로 추가
LottoIssuer 클래스는 다음과 같은 메서드와 생성자를 가집니다.
- 발행한 로또를 저장하는 lotteries
- 로또를 발행하는 issue 메서드
- 발행한 로또를 출력하는 print 메서드
